move get_latest_transaction to account
